QuickLIT-master/scopus/requirements.txt
flask==2.0.1
requests==2.31.0
python-dotenv==0.19.0
flask-cors==3.0.10
firebase-admin==6.8.0
python-dateutil==2.8.2 